requirements.txt
# Core ADK and Models
google-adk>=1.0
google-genai>=0.2.0

# Server
fastapi>=0.110.0
uvicorn>=0.29.0
python-telegram-bot>=21.0.1

# GCP Services
google-cloud-firestore>=2.15.0
google-cloud-dlp>=3.16.0
google-cloud-secret-manager>=2.18.0
google-cloud-logging>=3.9.0
google-cloud-trace>=1.11.2

# Config
pydantic-settings>=2.2.1

# Testing & Dev
pytest>=8.1.1
pytest-cov>=5.0.0
pytest-mock>=3.14.0
ruff>=0.3.0
mypy>=1.9.0
httpx>=0.27.0
